moved methods to single file
